refactor(components): change radius types to u8
Updates various components and functions to use `u8` for radius values instead of `i32`, ensuring consistency and reducing unnecessary size for values that are inherently non-negative.

Adds validation for level range parameters and refactors default detection range logic for clarity and reusability.

Removes unused plugin assignment in `main.rs` to clean up the codebase.

+++ crates/echos_assets/src/entities/components.rs
@@ -31,11 +31,11 @@ impl TurnActorData {
 #[derive(Serialize, Deserialize, Debug, Clone, Reflect)]
 pub struct ViewShedData {
     /// Vision radius in tiles
-    pub radius: i32,
+    pub radius: u8,
 }
 
 impl ViewShedData {
-    pub fn new(radius: i32) -> Self { Self { radius } }
+    pub fn new(radius: u8) -> Self { Self { radius } }
 }
 
 // impl From<ViewShedData> for ViewShed {
